add parallelism
While par_bridge is a really cool, useful feature, I'm not entirely
sure that this was a good idea: without benchmarks, I'm not certain
that it's actually processing entries substantially faster than before.

Still, it's managed to push my machine's cpu consumption to 100%,
which I think is a first for this one. That's always fun.